move project/ project-set/ components
