+++ src/components/index.ts
@@ -57,19 +57,12 @@ export { NavTab } from './NavTab';
 export { NavTabs } from './NavTabs';
 export { Output } from './Output';
 export { OutputList } from './OutputList';
-export { ProjectCreateForm } from './ProjectCreateForm';
-export { ProjectEditForm } from './ProjectEditForm';
-export { ProjectFinishedTip } from './ProjectFinishedTip';
-export { ProjectImportFromLabelplusStatus } from './ProjectImportFromLabelplusStatus';
 export { ProjectItem } from './ProjectItem';
 export { ProjectList } from './ProjectList';
 export { ProjectSetCreateForm } from './ProjectSetCreateForm';
 export { ProjectSetEditForm } from './ProjectSetEditForm';
 export { ProjectSetList } from './ProjectSetList';
 export { ProjectSetSettingBase } from './ProjectSetSettingBase';
-export { ProjectSettingBase } from './ProjectSettingBase';
-export { ProjectSettingTarget } from './ProjectSettingTarget';
-export { ProjectTargetList } from './ProjectTargetList';
 export { RoleRadioGroup } from './RoleRadioGroup';
 export { RoleSelect } from './RoleSelect';
 export { Spin } from './Spin';
@@ -90,4 +83,3 @@ export { UserEmailEditForm } from './UserEmailEditForm';
 export { UserInvitationList } from './UserInvitationList';
 export { UserPasswordEditForm } from './UserPasswordEditForm';
 export { VCodeInput } from './VCodeInput';
-export { ProjectImportForm } from './ProjectImportForm';
