+++ src/components/project/ProjectSettingTarget.tsx
@@ -5,7 +5,7 @@ import classNames from 'classnames';
 import React, { useState } from 'react';
 import { useIntl } from 'react-intl';
 import { useDispatch, useSelector } from 'react-redux';
-import api, { resultTypes } from '../apis';
+import api, { resultTypes } from '@/apis';
 import {
   EmptyTip,
   Icon,
@@ -14,12 +14,12 @@ import {
   ListItem,
   LIST_ITEM_DEFAULT_HEIGHT,
   Spin,
-} from '../components';
-import { FC, Target, Project } from '../interfaces';
-import { AppState } from '../store';
-import { increaseCurrentProjectTargetCount } from '../store/project/slice';
-import style from '../style';
-import { toLowerCamelCase } from '../utils';
+} from '@/components';
+import { FC, Target, Project } from '@/interfaces';
+import { AppState } from '@/store';
+import { increaseCurrentProjectTargetCount } from '@/store/project/slice';
+import style from '@/style';
+import { toLowerCamelCase } from '@/utils';
 
 /** 项目目标设置区域的属性接口 */
 interface ProjectSettingTargetProps {
